fix(#303): smugglr review nits -- chronological RFC3339 compare, round_f64 type
Addresses two of three non-blocking nits from smugglr's rust-idioms review
of PR #304 (signal 019db754-4507). Leaving `eprintln!` in `newest_age`
as-is because it matches the house convention across 14 files
(sync_actor.rs 10, mcp.rs 15, watch.rs 23, cluster.rs 6, ...); converting
mesh alone would be inconsistent. When/if the codebase adopts tracing
wholesale, mesh converts with it.

1. `newest_sampled_at` now parses DateTimes and compares chronologically
   instead of lex-comparing RFC3339 strings. Matters for cluster sync:
   peer-written rows with `Z` suffix and local rows with `+00:00` would
   silently diverge from chronological order under lex compare
   (`+`=0x2B < `Z`=0x5A). Falls back to string compare only when both
   sides fail to parse. New test pins both directions.

2. `round_f64(v, digits: u32)` -> `round_f64(v, digits: i32)`. Drops the
   pointless `as i32` cast at the call to `powi`. Literal `1` callers
   unchanged.

+++ src/main.rs
@@ -1964,8 +1964,8 @@ fn resolve_stale_cutoff() -> std::time::Duration {
     std::time::Duration::from_secs(secs)
 }
 
-fn round_f64(v: f64, digits: u32) -> f64 {
-    let factor = 10f64.powi(digits as i32);
+fn round_f64(v: f64, digits: i32) -> f64 {
+    let factor = 10f64.powi(digits);
     (v * factor).round() / factor
 }
